aider: Refactored the `get_mac_address_and_interface` function to use `ifaddresses()` instead of `gateways()` to get the default interface and its associated MAC address.
# Aider chat conversation:

USER: Traceback (most recent call last):
  File "/Users/djh/work/src/github.com/djh00t/module_klingon/klingon/utils.py", line 51, in <module>
    mac_address, interface = get_mac_address_and_interface()
  File "/Users/djh/work/src/github.com/djh00t/module_klingon/klingon/utils.py", line 35, in get_mac_address_and_interface
    gws = netifaces.gateways()
  File "/Users/djh/miniconda3/envs/klingon/lib/python3.9/site-packages/netifaces/__init__.py", line 189, in gateways
    raise NotImplementedError("No implementation for `gateways()` yet")
NotImplementedError: No implementation for `gateways()` yet

ASSISTANT: The error message indicates that the `gateways()` function from the `netifaces` module is not implemented. This function is used to get the default gateway and its associated network interface.

To fix this issue, we need to change the logic in the `get_mac_address_and_interface` function in the `klingon/utils.py` file. Instead of using the `gateways()` function, we can use the `ifaddresses()` function from the `netifaces` module to get the addresses associated with each network interface, and then check if the 'addr' field of the first address in the AF_INET family is not '127.0.0.1' (the loopback address). If it is not, we can consider this interface as the default interface.

Here is the *edit block* for the change:

```python
klingon/utils.py
<<<<<<< HEAD
        else:
            gws = netifaces.gateways()
            default_interface = gws['default'][netifaces.AF_INET][1]
            addresses = netifaces.ifaddresses(default_interface)
            mac_address = addresses[netifaces.AF_LINK][0]['addr']
            return mac_address, default_interface
=======
        else:
            for interface in netifaces.interfaces():
                addresses = netifaces.ifaddresses(interface)
                if netifaces.AF_INET in addresses:
                    if addresses[netifaces.AF_INET][0]['addr'] != '127.0.0.1':
                        mac_address = addresses[netifaces.AF_LINK][0]['addr']
                        return mac_address, interface
>>>>>>> updated
```