aider: Fixed KeyError when accessing MAC address of network interface.
# Aider chat conversation:

USER: Traceback (most recent call last):
  File "/Users/djh/work/src/github.com/djh00t/module_klingon/klingon/utils.py", line 52, in <module>
    mac_address, interface = get_mac_address_and_interface()
  File "/Users/djh/work/src/github.com/djh00t/module_klingon/klingon/utils.py", line 39, in get_mac_address_and_interface
    mac_address = addresses[netifaces.AF_LINK][0]['addr']
KeyError: -1000

ASSISTANT: The error message indicates that the `KeyError: -1000` is being raised. This error is raised when a dictionary is accessed with a key that does not exist in the dictionary. In this case, the key is `-1000`, which corresponds to `netifaces.AF_LINK`. This suggests that the `addresses` dictionary does not contain an entry for `netifaces.AF_LINK`.

This error can occur if the network interface does not have a MAC address, which can be the case for certain types of interfaces (e.g., loopback, tunnel, or virtual interfaces).

To fix this issue, we need to add a check to ensure that `netifaces.AF_LINK` exists in the `addresses` dictionary before trying to access it. Here is the *edit block* for the change:

```python
klingon/utils.py
<<<<<<< HEAD
        else:
            for interface in netifaces.interfaces():
                addresses = netifaces.ifaddresses(interface)
                if netifaces.AF_INET in addresses:
                    if addresses[netifaces.AF_INET][0]['addr'] != '127.0.0.1':
                        mac_address = addresses[netifaces.AF_LINK][0]['addr']
                        return mac_address, interface
=======
        else:
            for interface in netifaces.interfaces():
                addresses = netifaces.ifaddresses(interface)
                if netifaces.AF_INET in addresses and netifaces.AF_LINK in addresses:
                    if addresses[netifaces.AF_INET][0]['addr'] != '127.0.0.1':
                        mac_address = addresses[netifaces.AF_LINK][0]['addr']
                        return mac_address, interface
>>>>>>> updated
```

This change will ensure that the `KeyError` is not raised when trying to access the MAC address of a network interface.