aider: Fix KeyError when accessing MAC address of a network interface.
# Aider chat conversation:

ASSISTANT: The error message indicates that the `KeyError: -1000` is being raised. This error is raised when a dictionary is accessed with a key that does not exist in the dictionary. In this case, the key is `-1000`, which corresponds to `netifaces.AF_LINK`. This suggests that the `ifaddresses(interface)` dictionary does not contain an entry for `netifaces.AF_LINK`.

This error can occur if the network interface does not have a MAC address, which can be the case for certain types of interfaces (e.g., loopback, tunnel, or virtual interfaces).

To fix this issue, we need to add a check to ensure that `netifaces.AF_LINK` exists in the `ifaddresses(interface)` dictionary before trying to access it. Here is the *edit block* for the change:

```python
klingon/utils.py
<<<<<<< HEAD
    # Find the primary network interface
    primary_interface = None
    for interface in interfaces:
        if netifaces.ifaddresses(interface)[netifaces.AF_LINK][0]['addr'] == uuid.getnode():
            primary_interface = interface
            break
=======
    # Find the primary network interface
    primary_interface = None
    for interface in interfaces:
        addresses = netifaces.ifaddresses(interface)
        if netifaces.AF_LINK in addresses and addresses[netifaces.AF_LINK][0]['addr'] == uuid.getnode():
            primary_interface = interface
            break
>>>>>>> updated
```

This change will ensure that the `KeyError` is not raised when trying to access the MAC address of a network interface.